requirements.txt
opencv-contrib-python
opencv-python
numpy
pillow
